+++ elixir-generate-dockerfile/src/lib/app_config.ex
@@ -66,7 +66,8 @@ defmodule GenerateDockerfile.AppConfig do
       workspace_dir = Keyword.fetch!(args, :workspace_dir)
       default_erlang_version = Keyword.fetch!(args, :default_erlang_version)
       default_elixir_version = Keyword.fetch!(args, :default_elixir_version)
-      data = build_data(workspace_dir, default_erlang_version, default_elixir_version)
+      old_distillery_elixir_version = Keyword.fetch!(args, :old_distillery_elixir_version)
+      data = build_data(workspace_dir, default_erlang_version, default_elixir_version, old_distillery_elixir_version)
       {:ok, data}
     catch
       {:usage_error, message} -> {:ok, %{error: message}}
@@ -77,7 +78,7 @@ defmodule GenerateDockerfile.AppConfig do
     {:reply, Map.fetch(data, key), data}
   end
 
-  defp build_data(workspace_dir, default_erlang_version, default_elixir_version) do
+  defp build_data(workspace_dir, default_erlang_version, default_elixir_version, old_distillery_elixir_version) do
     project_id = get_project()
     project_id_for_display = project_id || "(unknown)"
     project_id_for_example = project_id || "my-project-id"
@@ -87,6 +88,7 @@ defmodule GenerateDockerfile.AppConfig do
     phoenix_version = Map.get(deps_info, :phoenix)
     phoenix_prefix = get_phoenix_prefix(phoenix_version)
 
+    default_elixir_version = adjust_default_elixir_version(default_elixir_version, old_distillery_elixir_version, distillery_version)
     {erlang_version, elixir_version} =
       get_tool_versions(workspace_dir, default_erlang_version, default_elixir_version)
 
@@ -186,6 +188,18 @@ defmodule GenerateDockerfile.AppConfig do
 
   defp get_phoenix_prefix(version) do
     if Version.compare(version, "1.3.0") == :lt, do: "phoenix", else: "phx"
+  end
+
+  defp adjust_default_elixir_version(default_elixir_version, _old_distillery_elixir_version, nil) do
+    default_elixir_version
+  end
+
+  defp adjust_default_elixir_version(default_elixir_version, old_distillery_elixir_version, distillery_version) do
+    if Version.compare(distillery_version, "2.1.0") == :lt do
+      old_distillery_elixir_version
+    else
+      default_elixir_version
+    end
   end
 
   defp get_tool_versions(workspace_dir, default_erlang_version, default_elixir_version) do